requirements.txt
wheel
telepot
pytz
pyrebase4